Refactor: extract PinnedMemoryBuffer
